Change: Move expected #31
